Back/back/requirements.txt
django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
django-filter>=23.0
reportlab>=4.0
Pillow>=10.0
pyotp>=2.9
qrcode[pil]>=7.4
python-decouple>=3.8
